web_service/requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.6
colorama==0.4.6
Django==3.2.20
djangorestframework==3.14.0
gunicorn==21.2.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.25.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
